fix: добавит отдельные модели для offers и destinations
